backend/requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.4
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-templated-mail==1.1.1
django4-background-tasks==1.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
et-xmlfile==1.1.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
lxml==5.1.0
msgpack==1.0.8
mysqlclient==2.2.1
num2words==0.5.13
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pygc==1.3.0
PyJWT==2.8.0
pykml==0.2.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
pyproj==3.6.1
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.6
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.1
zope.interface==6.4.post2
